Attach listener to add card, update styles font and colours
