+++ app/src/main/res/drawable/edit_text_bg.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="#E8F8FD" />
+    <solid android:color="@color/course_card_grey" />
 
     <corners
         android:radius="10dp" />
