+++ app/src/main/res/layout/activity_edit_card_page.xml
@@ -28,11 +28,11 @@
                 android:layout_marginStart="10dp"
                 android:layout_marginTop="5dp"
                 android:layout_weight="1"
-                android:fontFamily="sans-serif-medium"
+                android:fontFamily="@font/manrope_bold"
                 android:paddingTop="10dp"
                 android:text="Title"
-                android:textColor="#000000"
-                android:textSize="34sp"
+                android:textColor="@color/dark_grey"
+                android:textSize="24sp"
                 android:textStyle="bold" />
 
             <FrameLayout
@@ -69,10 +69,11 @@
                 android:layout_marginTop="10dp"
                 android:layout_weight="1"
                 android:background="@drawable/submit_button_bg"
-                android:fontFamily="sans-serif-medium"
+                android:fontFamily="@font/manrope_bold"
                 android:gravity="center"
                 android:padding="10dp"
                 android:text="Submit"
+                android:textAllCaps="false"
                 android:textColor="#FFFFFF"
                 android:textSize="16sp" />
