+++ app/src/main/res/layout/fill_in_card.xml
@@ -12,11 +12,11 @@
         android:layout_marginHorizontal="10dp"
         android:layout_marginTop="5dp"
         android:layout_weight="1"
-        android:fontFamily="@font/font"
+        android:fontFamily="@font/manrope_medium"
         android:gravity="center_vertical"
         android:includeFontPadding="false"
         android:text="Edit / Add New Card"
-        android:textColor="#000000"
+        android:textColor="@color/dark_grey"
         android:textSize="16sp"
         android:textStyle="bold" />
 
@@ -26,7 +26,7 @@
         android:layout_height="wrap_content"
         android:layout_marginHorizontal="10dp"
         android:layout_weight="20"
-        android:background="@drawable/flair_bg"
+        android:background="@drawable/edit_text_bg"
         android:ems="10"
         android:fontFamily="@font/manrope_medium"
         android:gravity="top|left"
@@ -43,7 +43,7 @@
         android:layout_marginHorizontal="10dp"
         android:layout_marginTop="5dp"
         android:layout_weight="1"
-        android:fontFamily="sans-serif-medium"
+        android:fontFamily="@font/manrope_medium"
         android:gravity="center_vertical"
         android:includeFontPadding="false"
         android:text="20 / 25 Character Limit"
