+++ build.gradle
@@ -5,7 +5,7 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:7.2.0-beta04'
+        classpath 'com.android.tools.build:gradle:7.3.0-alpha07'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.10"
 
         // Firebase
